--- src/components/keyboard/index.ts
@@ -1,2 +1,0 @@
-import { Keyboard } from './keyboard';
-export { Keyboard };
