+++ src/components/keypad/index.ts
@@ -1,0 +1,2 @@
+import { Keypad } from './keypad';
+export { Keypad };
